--- src/main/java/ningenme/net/api/domain/repository/mysql/ApplicationMetaRepository.java
@@ -1,10 +1,0 @@
-package ningenme.net.api.domain.repository.mysql;
-
-import lombok.NonNull;
-import ningenme.net.api.domain.entity.ApplicationMeta;
-import ningenme.net.api.domain.value.ApplicationMetaId;
-
-public interface ApplicationMetaRepository {
-  ApplicationMeta getLatestOne(@NonNull final ApplicationMetaId applicationMetaId);
-  void post(@NonNull final ApplicationMetaId applicationMetaId);
-}
