--- src/main/java/ningenme/net/api/domain/repository/mysql/NetUserRepository.java
@@ -1,8 +1,0 @@
-package ningenme.net.api.domain.repository.mysql;
-
-import lombok.NonNull;
-import ningenme.net.api.domain.entity.NetUser;
-
-public interface NetUserRepository {
-  void post(@NonNull final NetUser netUser);
-}
