+++ src/main/java/ningenme/net/api/domain/service/ApplicationMetaService.java
@@ -1,10 +1,23 @@
 package ningenme.net.api.domain.service;
 
 import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import ningenme.net.api.domain.entity.ApplicationMeta;
 import ningenme.net.api.domain.value.ApplicationMetaId;
+import ningenme.net.api.infrastructure.mysql.ApplicationMetaRepository;
+import org.springframework.stereotype.Service;
 
-public interface ApplicationMetaService {
-  ApplicationMeta getLatestOne(@NonNull final ApplicationMetaId applicationMetaId);
-  void post(@NonNull final ApplicationMetaId applicationMetaId);
+@Service
+@RequiredArgsConstructor
+public class ApplicationMetaService {
+
+  private final ApplicationMetaRepository applicationMetaRepository;
+
+  public ApplicationMeta getLatestOne(@NonNull final ApplicationMetaId applicationMetaId) {
+    return applicationMetaRepository.getLatestOne(applicationMetaId);
+  }
+
+  public void post(@NonNull ApplicationMetaId applicationMetaId) {
+    applicationMetaRepository.post(applicationMetaId);
+  }
 }
