+++ src/main/java/ningenme/net/api/domain/service/NetUserService.java
@@ -1,8 +1,17 @@
 package ningenme.net.api.domain.service;
 
 import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import ningenme.net.api.domain.entity.NetUser;
+import ningenme.net.api.infrastructure.mysql.NetUserRepository;
+import org.springframework.stereotype.Service;
 
-public interface NetUserService {
-  void post(@NonNull final NetUser netUser);
+@Service
+@RequiredArgsConstructor
+public class NetUserService {
+  private final NetUserRepository netUserRepository;
+
+  public void post(@NonNull final NetUser netUser) {
+    netUserRepository.post(netUser);
+  }
 }
