+++ src/main/java/ningenme/net/api/infrastructure/mysql/ApplicationMetaRepository.java
@@ -3,24 +3,21 @@ package ningenme.net.api.infrastructure.mysql;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import ningenme.net.api.domain.entity.ApplicationMeta;
-import ningenme.net.api.domain.repository.mysql.ApplicationMetaRepository;
 import ningenme.net.api.domain.value.ApplicationMetaId;
 import ningenme.net.api.infrastructure.mysql.mapper.ApplicationMetaMapper;
 import org.springframework.stereotype.Repository;
 
 @RequiredArgsConstructor
 @Repository
-public class ApplicationMetaRepositoryImpl implements ApplicationMetaRepository {
+public class ApplicationMetaRepository {
 
   private final ApplicationMetaMapper applicationMetaMapper;
 
-  @Override
   public ApplicationMeta getLatestOne(@NonNull final ApplicationMetaId applicationMetaId) {
     return ApplicationMeta.of(
             applicationMetaMapper.selectLatestOne(applicationMetaId.getValue()));
   }
 
-  @Override
   public void post(@NonNull final ApplicationMetaId applicationMetaId) {
     applicationMetaMapper.insert(applicationMetaId.getValue());
   }
